Reorganize components by how they’re outputted
This commit moves a few PL files around to match the way Fractal displays components.

For example, “button—dropdown-state” was in the “02-molecules” folder. But its parent component, “button,” is in “01-atoms”, so Fractal displays “Button Dropdown State” under “Atoms.”

So in this commit, we’ve moved “button—dropdown-state” to “01-atoms.”